+++ app/validators/local/unique-attributes.js
@@ -1,0 +1,2 @@
+import uniqueAttributes from 'ember-flexberry/validators/local/unique-attributes';
+export default uniqueAttributes;