+++ tests/acceptance/components/flexberry-objectlistview/filther/folv-filter-on-toolbar-test.js
@@ -20,8 +20,8 @@ executeTest('check filter on toolbar with filter projection', (store, assert, ap
     let sp2 = new SimplePredicate('type.name', FilterOperator.Neq, '');
     let cp = new ComplexPredicate(Condition.And, sp1, sp2);
 
-    let builder2 = new Query.Builder(store).from(modelName).selectByProjection('SuggestionL').where(cp).top(1);
-    store.query(modelName, builder2.build()).then((result) => {
+    let builder = new Query.Builder(store).from(modelName).selectByProjection('SuggestionL').where(cp).top(1);
+    store.query(modelName, builder.build()).then((result) => {
       let arr = result.toArray();
       adressEtalone = arr.objectAt(0).get('address');
       typeNameEtalone = arr.objectAt(0).get('type.name');
@@ -36,7 +36,7 @@ executeTest('check filter on toolbar with filter projection', (store, assert, ap
       Ember.run(() => {
         fillIn($filterInput, adressEtalone);
       });
-      
+
       let refreshFunction =  function() {
         $filterApplyButton.click();
       };
@@ -49,7 +49,7 @@ executeTest('check filter on toolbar with filter projection', (store, assert, ap
         let currentModel = Ember.get(controller, 'model');
         let filteredByCommonProjectionCountN = Ember.get(currentModel, 'meta.count');
         assert.ok(filteredByCommonProjectionCountN >= 1, `Found ${filteredByCommonProjectionCountN} records by common projection filtered by "${adressEtalone}".`);
-        
+
         // 2) Filter by type.name as typeNameEtalone by filter projection containing only type.name property and get at least 1 record.
         Ember.run(() => {
           Ember.set(controller, 'filterProjectionName', 'TestFilterOnToolbarView');
@@ -58,7 +58,7 @@ executeTest('check filter on toolbar with filter projection', (store, assert, ap
         Ember.run(() => {
           fillIn($filterInput, typeNameEtalone);
         });
-        
+
         let done2 = assert.async();
         refreshListByFunction(refreshFunction, controller).then(($list) => {
           let currentModel = Ember.get(controller, 'model');
@@ -75,7 +75,7 @@ executeTest('check filter on toolbar with filter projection', (store, assert, ap
             let currentModel = Ember.get(controller, 'model');
             let filteredByFilterProjectionCount2 = Ember.get(currentModel, 'meta.count');
             assert.ok(filteredByCommonProjectionCountN > filteredByFilterProjectionCount2, `Found ${filteredByFilterProjectionCount2} records by filter projection filtered by "${adressEtalone}".`);
-            
+
             Ember.run(() => {
               Ember.set(controller, 'filterProjectionName', undefined);
             });