+++ tests/dummy/app/router.js
@@ -186,6 +186,7 @@ Router.map(function() {
 
   // Components acceptance tests forms.
   this.route('components-acceptance-tests/flexberry-lookup/settings-example-autofill-by-limit');
+  this.route('components-acceptance-tests/flexberry-lookup/settings-example-custom-window');
   this.route('components-acceptance-tests/flexberry-lookup/base-operations');
   this.route('components-acceptance-tests/flexberry-lookup/settings-example');
   this.route('components-acceptance-tests/flexberry-lookup/settings-example-autocomplete');